requirements.txt
discord.py<=3.0.0
